split up board.h into seperate files
