+++ server/source/serverlib/game/board/taxfield.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <game/board/field.h>
+
+class TaxField : public Field
+{
+public:
+    TaxField(int amount);
+
+    int amount() const;
+
+protected:
+    int m_amount;
+};
