+++ server/source/serverlib/models/boardmodel.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <game/board.h>
+#include <game/board/board.h>
 #include <utils/staticstorage.h>
 
 class BoardModel : public StaticStorage<BoardModel>
